exerciocio 60 refeito codigo - treino
